% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oclo.R
\name{oclo}
\alias{oclo}
\title{Order Constrained Linear Optimization}
\usage{
oclo(x, ...)
}
\arguments{
\item{blah}{does blah things}
}
\description{
This function allows you to do silly things.
}
\examples{
oclo()
}
\keyword{oclo}

