% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oclo.R
\name{oclo}
\alias{oclo}
\title{Order Constrained Linear Optimization}
\usage{
oclo(x, ...)
}
\arguments{
\item{blah}{does blah things}
}
\description{
This function allows you to do silly things.
}
\examples{
# Generate some data
n <- 25
x1 <- rnorm(n)
x2 <- nrorm(n)
y <- 10 + 5*x1 + 2*x2 + rnorm(n)

fit.oclo <- oclo(y~x1+x2)

}
\keyword{oclo}

